--- api/src/main/kotlin/org/jetbrains/kotlinx/dl/datasets/Preprocessor.kt
@@ -1,14 +1,0 @@
-/*
- * Copyright 2020 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
- */
-
-package org.jetbrains.kotlinx.dl.datasets
-
-public interface Preprocessor {
-    public fun apply(data: FloatArray): FloatArray // move to shape
-}
-
-
-
-
