+++ api/src/main/kotlin/org/jetbrains/kotlinx/dl/datasets/preprocessors/Rescaling.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.kotlinx.dl.datasets
+package org.jetbrains.kotlinx.dl.datasets.preprocessors
 
 public class Rescaling(public var scalingCoefficient: Float = 255f) : Preprocessor {
     override fun apply(data: FloatArray): FloatArray {
