+++ api/src/main/kotlin/org/jetbrains/kotlinx/dl/datasets/preprocessors/image/InterpolationType.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.kotlinx.dl.datasets
+package org.jetbrains.kotlinx.dl.datasets.preprocessors.image
 
 public enum class InterpolationType {
     BILINEAR, NEAREST, BICUBIC
