+++ api/src/main/kotlin/org/jetbrains/kotlinx/dl/datasets/preprocessors/image/Resize.kt
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2020 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlinx.dl.datasets.preprocessors.image
+
+import org.jetbrains.kotlinx.dl.datasets.preprocessors.ImageShape
+import java.awt.RenderingHints
+import java.awt.image.BufferedImage
+
+public class Resize(
+    public var outputWidth: Int = 100,
+    public var outputHeight: Int = 100,
+    public var interpolation: InterpolationType = InterpolationType.BILINEAR,
+    public var renderingSpeed: RenderingSpeed = RenderingSpeed.MEDIUM,
+    public var enableAntialiasing: Boolean = true
+) : ImagePreprocessor {
+    override fun apply(image: BufferedImage, inputShape: ImageShape): Pair<BufferedImage, ImageShape> {
+        val resizedImage = BufferedImage(outputWidth, outputHeight, BufferedImage.TYPE_3BYTE_BGR)
+        val graphics2D = resizedImage.createGraphics()
+
+        val renderingHint = when (interpolation) {
+            InterpolationType.BILINEAR -> RenderingHints.VALUE_INTERPOLATION_BILINEAR
+            InterpolationType.NEAREST -> RenderingHints.VALUE_INTERPOLATION_NEAREST_NEIGHBOR
+            InterpolationType.BICUBIC -> RenderingHints.VALUE_INTERPOLATION_BICUBIC
+        }
+
+        val renderingSpeed = when (renderingSpeed) {
+            RenderingSpeed.FAST -> RenderingHints.VALUE_RENDER_SPEED
+            RenderingSpeed.SLOW -> RenderingHints.VALUE_RENDER_QUALITY
+            RenderingSpeed.MEDIUM -> RenderingHints.VALUE_RENDER_DEFAULT
+        }
+
+        graphics2D.setRenderingHint(RenderingHints.KEY_INTERPOLATION, renderingHint)
+        graphics2D.setRenderingHint(RenderingHints.KEY_RENDERING, renderingSpeed)
+
+        if (enableAntialiasing) {
+            graphics2D.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON)
+        }
+
+        graphics2D.drawImage(image, 0, 0, outputWidth, outputHeight, null)
+        graphics2D.dispose()
+
+        return Pair(
+            resizedImage,
+            ImageShape(outputWidth.toLong(), outputHeight.toLong(), channels = inputShape.channels)
+        )
+    }
+}
