+++ api/src/main/kotlin/org/jetbrains/kotlinx/dl/datasets/preprocessors/image/Rotate.kt
@@ -3,8 +3,9 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.kotlinx.dl.datasets
+package org.jetbrains.kotlinx.dl.datasets.preprocessors.image
 
+import org.jetbrains.kotlinx.dl.datasets.preprocessors.ImageShape
 import java.awt.RenderingHints
 import java.awt.geom.AffineTransform
 import java.awt.image.BufferedImage
